Allow to override the remote name
Currently we hard-code origin as the remote to use for git.
Allow the user to override this via TARGET_ORIGIN variable (which
defaults to 'origin' like before)

Tested with:
make upgrade
make TARGET_ORIGIN=fork upgrade

And the commands used the 'origin' remote and the 'fork' one
respectively.

Reason for this is that I hate calling my forks 'origin' and I rarely
even have an 'origin' remote. I call my remotes 'fork' so I do not push
branches upstream accidentally.