Make the default policy for hub/* secrets more restrictive
Leave only the read one for the time being as it seems to be a saner
default. If more granularity is needed customizations are qreuired in
any case.

Tested and I can still correctly access the secret on the config-demo
app on the regional cluster.

This is the ansible port of ed592b33ba044b579b4d92f448d5abbc493dc447