Support k8s 1.24
In kubernetes 1.24 (OCP 4.11) the token for service account does not get
created anymore [1]. In order to support all of our supported OCP versions
with the least effort, we just create a secret and associate it to the
golang-external-secret service account.

While we're at it we simplify the ansible a code a bit and simply always
rely on having the golang-external-secrets secret created by default.

Tested on OCP 4.8, 4.9, 4.10 and 4.11. In all tests I saw the
config-demo-secret being created inside the cluster and 'make install'
succeeded correctly.

Closes: https://issues.redhat.com/browse/MBP-321

[1] https://github.com/kubernetes/kubernetes/blob/master/CHANGELOG/CHANGELOG-1.24.md#deprecation=