Add a check a check_missing_secrets parameter to vault_load_secrets
When this is set to true we read the 'values_secret_template' parameter,
parse the yaml file pointed to it and make sure that al the keys in
there exist in the values-secret.yaml file.

The way the implementation works is by flattening the dictionary like
the following. This is the starting yaml/dictionary:

    secrets:
      config-demo:
	secret: demo123
      cluster_alejandro:
	name: alejandro
	bearerToken: foo
      test:
	s3.accessKey: "1234"
	s3.secretKey: "4321"
    files:
      cluster_alejandro_ca: /home/michele/ca.crt

This will generate the following flattended structure:

    {'secrets.config-demo.secret': 'demo123',
    'secrets.cluster_alejandro.name': 'alejandro',
    'secrets.cluster_alejandro.bearerToken': 'foo',
    'secrets.test.s3.accessKey': '1234',
    'secrets.test.s3.secretKey': '4321',
    'files.cluster_alejandro_ca': '/home/michele/ca.crt'}

We then take the keys of the flattened structure and compare them
as sets. If the keys of the values-secret.yaml.template file are a
subset of those in ~/values-secret.yaml it means all of them are
present.