Add all the code to actually verify secrets against values-secret-template.yaml but keep it disabled
This adds the last piece to actually be able to flip the check_missing_secrets
switch to True.

So as is this is still not being run. The idea here is to socialize this
feature a bit more and then flip it one. Once set to True, the
'values-secret-template.yaml' becomes a *required* part of the pattern.