Add support for private repos
This needs the corresponding PR from the operator
(https://www.github.com/validatedpatterns/patterns-operator/pull/139).
The way it works is that if "global.privateRepo" is set to true, we
add an acm policy on the hub only that reads the secret from the openshift-gitops
namespaces and copies it to the open-cluster-manager. And then we use
another policy that pushes the secret just copied to
open-cluster-managemnt to the openshift-gitops + pattern-name-group-one namespaces
so that the two argo instances can consume the private repositories.

Tested end to end with both https and ssh private repository.