Add overrides for MCE IIB
We need to be able to explicitely override the Source and the Channel
used to install the MCE multicluster-engine operator. The reason for
this is that when we install ACM from an IIB, the default channel that
will be used for MCE won't be present in the IIB, so we need to be able
to be flexible in determining where the MCE gets installed from
(sourcecatalog and channel)

We default to redhat-operators because otherwise when installing gitops
via IIB the ACM operator finds a channel for MCE that does not exist, so
we force the source all the time.