Fix pki bind mount when using podman machine
Currently we test for the existance of local TLS folder and bind mount
them in containers. This does not work correctly when using podman
machine, because a the test is running on the host and so we might bind
mount the wrong folder. For example:
On Mac OSX /etc/pki does not exist on the folder and so we bind mount
/etc/ssl even though /etc/pki does exist in the podman machine VM

Co-Authored-By: Ákos Erős <aeros@redhat.com>